Rename ContentFacade to ContentManager
