+++ Resources/config/services.xml
@@ -116,8 +116,8 @@
             <tag name="kernel.event_subscriber"/>
         </service>
 
-        <!-- ContentFacade -->
-        <service id="sulu_content.content_facade" class="Sulu\Bundle\ContentBundle\Content\Application\ContentFacade\ContentFacade">
+        <!-- Content Manager -->
+        <service id="sulu_content.content_manager" class="Sulu\Bundle\ContentBundle\Content\Application\ContentManager\ContentManager">
             <argument type="service" id="sulu_content.content_resolver"/>
             <argument type="service" id="sulu_content.content_persister"/>
             <argument type="service" id="sulu_content.content_projection_normalizer"/>
@@ -125,6 +125,6 @@
             <argument type="service" id="sulu_content.content_workflow"/>
         </service>
 
-        <service id="Sulu\Bundle\ContentBundle\Content\Application\ContentFacade\ContentFacadeInterface" alias="sulu_content.content_facade"/>
+        <service id="Sulu\Bundle\ContentBundle\Content\Application\ContentManager\ContentManagerInterface" alias="sulu_content.content_manager"/>
     </services>
 </container>
