+++ Tests/Application/ExampleTestBundle/Resources/config/services.yaml
@@ -19,7 +19,7 @@ services:
             - '@sulu_core.list_builder.field_descriptor_factory'
             - '@sulu_core.doctrine_list_builder_factory'
             - '@sulu_core.doctrine_rest_helper'
-            - '@sulu_content.content_facade'
+            - '@sulu_content.content_manager'
             - '@doctrine.orm.entity_manager'
         tags:
             - { name: sulu.context, context: admin }
